+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 0.49.0 (Unreleased)
+
+ENHANCEMENTS:
+
+* mdb: add `web_sql` attribute to `config.access` entity in `yandex_mdb_postgresql_cluster` resource and data source
+* mdb: add `performance_diagnostics` attribute to `config` entity in `yandex_mdb_postgresql_cluster` resource and data source
+
 ## 0.48.0 (December 22, 2020)
 
 BUG FIXES: